Issue #24, analyze over indexing refacotring
